+++ monopoly/src/main/java/org/centrale/projet/monopoly/Case.java
@@ -1,27 +1,22 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package edu.centralenantes.projetmedevei2;
-
+package org.centrale.projet.monopoly;
 
 /**
  *
  * @author louis-alexandre
  */
 public abstract class Case {
+
     private String nom;
 
     public Case(String nom) {
         this.nom = nom;
     }
-    
-    public Case(Case c){
+
+    public Case(Case c) {
         this.nom = c.nom;
     }
-    
-    public Case(){
+
+    public Case() {
         this.nom = "Départ";
     }
 
@@ -32,11 +27,11 @@ public abstract class Case {
     public void setNom(String nom) {
         this.nom = nom;
     }
-    
+
     public abstract String toString();
-    
-    public boolean equals(Case c){
-        return this.getNom()==c.getNom();
+
+    public boolean equals(Case c) {
+        return this.getNom() == c.getNom();
     }
-    
+
 }
