+++ monopoly/src/main/java/org/centrale/projet/monopoly/Depart.java
@@ -1,0 +1,9 @@
+package org.centrale.projet.monopoly;
+
+/**
+ *
+ * @author louis-alexandre
+ */
+public class Depart {
+    
+}
