+++ monopoly/src/main/java/org/centrale/projet/monopoly/NonAchetable.java
@@ -1,0 +1,14 @@
+package org.centrale.projet.monopoly;
+
+/**
+ *
+ * @author louis-alexandre
+ */
+public class NonAchetable extends Case {
+
+    @Override
+    public String toString() {
+        return this.getNom();
+    }
+
+}
